% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{extract_recipient_details}
\alias{extract_recipient_details}
\title{extract recipient details from invoice}
\usage{
extract_recipient_details(path, regex = regex_pattern())
}
\arguments{
\item{path}{\code{string} path of invoice pdf}
}
\description{
extract recipient details from invoice
}
