% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{invoice_folder_to_adress_book}
\alias{invoice_folder_to_adress_book}
\title{convert invoice folder to adress book}
\usage{
invoice_folder_to_adress_book(dir)
}
\arguments{
\item{dir}{\code{string} path of invoice folder}
}
\description{
crawl through invoice folder, parse invoices,
    extract recipient and combine to collect address book
}
